恢复被 a66a4c1 误回滚的三个关键修复
a66a4c1 (大改基线2) 把以下修复全部回滚，导致本地与云端 85e80b5 不一致：
- graph_transformer.py: AMP dtype 修复丢失（开 use_amp 会崩）
- train_graph_model.py: 三个消融开关 + 互斥校验 + gcn_only 等深度丢失
- graph_builder.py: hybrid = distance+long_range 修复丢失（hybrid 退化为 distance）

本次从原修复 commit 恢复：
- graph_transformer.py ← 85e80b5
- train_graph_model.py ← a98d119
- graph_builder.py ← 2e4317a

恢复后本地与云端 85e80b5 的 Python 代码一致，default.yaml 为云端同步版。

+++ graph_transform/models/graph_transformer.py
@@ -585,6 +585,7 @@ class GraphTransformer(nn.Module):
             max_bonds = int(precomputed_max_bonds)
         else:
             max_bonds = int(bond_counts.max().item()) if batch_size > 0 else 0
+        # predictions 容器固定 float32；AMP 下 bond_logits 在赋值处用 .to() 对齐。
         predictions = torch.zeros(batch_size, max_bonds, device=node_features.device)
 
         if max_bonds > 0:
@@ -618,7 +619,9 @@ class GraphTransformer(nn.Module):
                 bond_feature_parts.append(h_src * h_dst)
             bond_features = torch.cat(bond_feature_parts, dim=-1)
             bond_logits_flat = self.bond_head(bond_features).squeeze(-1)
-            predictions[valid_bond_mask] = bond_logits_flat
+            # AMP 下 bond_head 输出为 half，predictions 容器为 float32，需对齐 dtype 后赋值。
+            # （autocast 不改 NodeEncoder 的 embedding 查表输出，故 node_features.dtype 不可靠）
+            predictions[valid_bond_mask] = bond_logits_flat.to(predictions.dtype)
         else:
             bond_logits_flat = torch.empty(0, device=node_features.device)
